# apps/power-platform/solution/Connector/itsup_ItSupportApi_openapidefinition.json
{
  "swagger": "2.0",
  "info": {
    "title": "IT Support API",
    "description": "The action/integration API for the IT support demo: retrieval-augmented answers over the knowledge base, profile lookup for the signed-in caller, and IT ticket creation. The TypeScript client, the C# server structure and the Power Platform custom connector all derive from this document.\n",
    "contact": {
      "name": "Ludvig Aldén"
    },
    "version": "0.1.0"
  },
  "basePath": "/api",
  "paths": {
    "/config": {
      "get": {
        "tags": [
          "it-support"
        ],
        "summary": "Get client configuration",
        "description": "Returns the client bootstrap configuration. Anonymous by design: it tells the SPA whether sign-in is configured and, if so, which Entra client and scope to use. When authentication is not configured the auth property is null and the client skips sign-in.\n",
        "operationId": "GetConfig",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "The client configuration.",
            "schema": {
              "$ref": "#/definitions/ClientConfig"
            }
          }
        },
        "x-ms-visibility": "internal"
      }
    },
    "/me": {
      "get": {
        "tags": [
          "it-support"
        ],
        "summary": "Get my profile",
        "description": "Returns the caller's profile. Sign-in is optional: a signed-in caller's profile is read from the directory on the caller's behalf, including the manager when one is recorded, and an anonymous caller receives the guest profile.\n",
        "operationId": "GetMyProfile",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "The caller's profile, or the guest profile when anonymous.",
            "schema": {
              "$ref": "#/definitions/UserProfile"
            }
          },
          "400": {
            "description": "The request was malformed."
          },
          "401": {
            "description": "Presented credentials failed authentication."
          },
          "503": {
            "description": "The signed-in caller's profile could not be read from the directory — for example when the directory on-behalf-of access still needs a one-time admin consent, or the directory is unreachable. Anonymous callers are unaffected: they receive the guest profile.\n",
            "schema": {
              "$ref": "#/definitions/ProfileUnavailable"
            }
          }
        }
      }
    },
    "/answers": {
      "post": {
        "tags": [
          "it-support"
        ],
        "summary": "Answer a question",
        "description": "Answers a question against the knowledge base and returns the answer text with citations, plus the retrieved chunk texts, which evaluation consumes. Anonymous callers are answered as guests. Requests are rate limited per client address; beyond the limit the endpoint answers 429.\n",
        "operationId": "AnswerQuestion",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AnswerRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The answer with citations and retrieved chunks.",
            "schema": {
              "$ref": "#/definitions/Answer"
            }
          },
          "400": {
            "description": "The request body was malformed."
          },
          "401": {
            "description": "Presented credentials failed authentication."
          },
          "429": {
            "description": "Too many requests from this client address."
          }
        }
      }
    },
    "/tickets": {
      "post": {
        "tags": [
          "it-support"
        ],
        "summary": "Create support ticket",
        "description": "Creates an IT ticket and returns it with its number. The caller is taken from the access token when one is presented and recorded as a guest otherwise — never from the request body. Requests are rate limited per client address; beyond the limit the endpoint answers 429.\n",
        "operationId": "CreateTicket",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTicketRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The ticket was created.",
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "400": {
            "description": "The request body was malformed."
          },
          "401": {
            "description": "Presented credentials failed authentication."
          },
          "429": {
            "description": "Too many requests from this client address."
          }
        }
      }
    }
  },
  "definitions": {
    "Person": {
      "type": "object",
      "description": "A directory person.",
      "required": [
        "displayName",
        "email"
      ],
      "properties": {
        "displayName": {
          "type": "string",
          "description": "The person's display name."
        },
        "email": {
          "type": "string",
          "description": "The person's email address.",
          "format": "email"
        }
      }
    },
    "AuthConfig": {
      "type": "object",
      "description": "Entra configuration for the client's sign-in.",
      "required": [
        "clientId",
        "tenantId",
        "scope"
      ],
      "properties": {
        "clientId": {
          "type": "string",
          "description": "The client application (client) ID."
        },
        "tenantId": {
          "type": "string",
          "description": "The directory (tenant) ID."
        },
        "scope": {
          "type": "string",
          "description": "The scope to request for API access."
        }
      }
    },
    "ClientConfig": {
      "type": "object",
      "description": "Client bootstrap configuration.",
      "required": [
        "auth"
      ],
      "properties": {
        "auth": {
          "description": "Sign-in configuration, or null when authentication is not configured.",
          "allOf": [
            {
              "$ref": "#/definitions/AuthConfig"
            }
          ]
        }
      }
    },
    "UserProfile": {
      "type": "object",
      "description": "The signed-in caller's profile.",
      "required": [
        "displayName",
        "email"
      ],
      "properties": {
        "displayName": {
          "type": "string",
          "description": "The caller's display name."
        },
        "email": {
          "type": "string",
          "description": "The caller's email address.",
          "format": "email"
        },
        "department": {
          "type": "string",
          "description": "The caller's department, when recorded.",
          "x-nullable": true
        },
        "manager": {
          "description": "The caller's manager, when one is recorded.",
          "allOf": [
            {
              "$ref": "#/definitions/Person"
            }
          ]
        }
      }
    },
    "ProfileUnavailable": {
      "type": "object",
      "description": "Why the signed-in caller's profile could not be read.",
      "required": [
        "code"
      ],
      "properties": {
        "code": {
          "type": "string",
          "description": "consent_required when the directory access needs a one-time admin consent; unavailable for any other directory failure.\n",
          "enum": [
            "consent_required",
            "unavailable"
          ]
        }
      }
    },
    "AnswerRequest": {
      "type": "object",
      "description": "A question for the knowledge base.",
      "required": [
        "question"
      ],
      "properties": {
        "question": {
          "type": "string",
          "description": "The question to answer.",
          "maxLength": 2000,
          "x-ms-summary": "Question"
        }
      }
    },
    "Citation": {
      "type": "object",
      "description": "A citation backing an answer.",
      "required": [
        "title",
        "url"
      ],
      "properties": {
        "title": {
          "type": "string",
          "description": "The cited article's title."
        },
        "url": {
          "type": "string",
          "description": "A link to the cited article."
        }
      }
    },
    "RetrievedChunk": {
      "type": "object",
      "description": "A retrieved knowledge-base chunk. Its text feeds evaluation.",
      "required": [
        "title",
        "content"
      ],
      "properties": {
        "title": {
          "type": "string",
          "description": "The chunk's article title."
        },
        "content": {
          "type": "string",
          "description": "The chunk's text."
        }
      }
    },
    "Answer": {
      "type": "object",
      "description": "An answer with its citations and retrieved chunks.",
      "required": [
        "text",
        "citations",
        "chunks"
      ],
      "properties": {
        "text": {
          "type": "string",
          "description": "The answer text."
        },
        "citations": {
          "type": "array",
          "description": "The citations backing the answer.",
          "items": {
            "$ref": "#/definitions/Citation"
          }
        },
        "chunks": {
          "type": "array",
          "description": "The retrieved chunks, for evaluation.",
          "items": {
            "$ref": "#/definitions/RetrievedChunk"
          }
        }
      }
    },
    "CreateTicketRequest": {
      "type": "object",
      "description": "A request to create an IT ticket.",
      "required": [
        "shortDescription"
      ],
      "properties": {
        "shortDescription": {
          "type": "string",
          "description": "A one-line summary of the problem.",
          "maxLength": 160,
          "x-ms-summary": "Short description"
        },
        "description": {
          "type": "string",
          "description": "A longer description of the problem.",
          "x-nullable": true,
          "x-ms-summary": "Description"
        }
      }
    },
    "Ticket": {
      "type": "object",
      "description": "A created IT ticket.",
      "required": [
        "number",
        "shortDescription",
        "caller",
        "createdAt"
      ],
      "properties": {
        "number": {
          "type": "string",
          "description": "The ticket number, on the form IT-yyyyMMdd-XXXX where XXXX is four characters from the Crockford base32 alphabet (pattern ^IT-\\d{8}-[0-9A-HJKMNP-TV-Z]{4}$).\n"
        },
        "url": {
          "type": "string",
          "description": "A link to the ticket in the ticketing system, when one exists.",
          "x-nullable": true
        },
        "shortDescription": {
          "type": "string",
          "description": "The ticket's one-line summary."
        },
        "caller": {
          "$ref": "#/definitions/Person"
        },
        "createdAt": {
          "type": "string",
          "description": "When the ticket was created.",
          "format": "date-time"
        }
      }
    }
  },
  "securityDefinitions": {
    "entra": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
      "tokenUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/token",
      "scopes": {
        "api://copaz-api/access_as_user": "Access the IT Support API on behalf of the signed-in user."
      },
      "description": "Microsoft Entra ID (authorization code flow with PKCE)."
    }
  },
  "security": [
    {
      "entra": [
        "api://copaz-api/access_as_user"
      ]
    }
  ],
  "tags": [
    {
      "name": "it-support",
      "description": "IT support operations for the assistant, the portal and the agent."
    }
  ],
  "host": "@environmentVariables('itsup_ApiHost')",
  "schemes": [
    "https"
  ]
}